separate dashboard handlers from index file for better editing
